+++ tests/apps/idempotency_add_unique_meta_app/migrations/0002_relatedtesttable_relatedtesttable_uniq.py
@@ -13,7 +13,6 @@ class Migration(migrations.Migration):
         migrations.AddConstraint(
             model_name="relatedtesttable",
             constraint=models.UniqueConstraint(
-                condition=models.Q(("test_field_int__gt", 0)),
                 fields=("test_field_int", "test_field_str"),
                 name="relatedtesttable_uniq",
             ),